backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gitpython==3.1.43
networkx==3.4.2
mistralai==2.4.0
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.27.2
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-typescript==0.23.2
pydantic==2.10.3
python-multipart==0.0.12
